# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CPI Artifacts API
  version: 1.0.0
  description: API Telegram Mini App для сбора материалов мероприятий.
servers:
  - url: https://artifacts.62-113-105-225.sslip.io
security:
  - sessionCookie: []
...
paths:
  /api/v1/auth/telegram:
    post:
      tags: [auth]
      security: []
      summary: Проверка подписанного Telegram initData и создание сессии
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [initData]
              properties:
                initData: { type: string, maxLength: 16384 }
      responses:
        '200':
          description: Сессия создана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Завершить сессию
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '204': { description: Сессия завершена }
...
  /api/v1/events:
    get:
      tags: [events]
      summary: Поиск и фильтрация опубликованных мероприятий
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 200 } }
        - { name: city, in: query, schema: { type: string } }
        - { name: format, in: query, schema: { $ref: '#/components/schemas/EventFormat' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/EventStatus' } }
        - { name: cursor, in: query, schema: { type: string, format: uuid } }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
          }
      responses:
        '200':
          description: Страница мероприятий
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Event' }
                  nextCursor: { type: [string, 'null'], format: uuid }
  /api/v1/events/{eventKey}:
    get:
      tags: [events]
      summary: Мероприятие по UUID, slug или короткому коду
      parameters:
        - { name: eventKey, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Мероприятие
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/events/{eventId}/submissions:
    get:
      tags: [submissions]
      summary: Свои отправки в мероприятие
      parameters:
        - { $ref: '#/components/parameters/EventId' }
      responses:
        '200': { description: Список отправок }
    post:
      tags: [submissions]
      summary: Идемпотентно создать отправку
      parameters:
        - { $ref: '#/components/parameters/EventId' }
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmissionInput' }
      responses:
        '201':
          description: Отправка создана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Submission' }
        '409': { description: Приём закрыт или профиль не заполнен }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cpi_artifacts_session
  parameters:
    CsrfToken:
      name: x-csrf-token
      in: header
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
    EventId:
      name: eventId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    EventFormat:
      enum: [offline, online, hybrid]
    EventStatus:
      enum: [draft, published, running, finished, archived]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        telegramUserId: { type: string, pattern: '^\d+$' }
        fullName: { type: [string, 'null'] }
        organization: { type: [string, 'null'] }
        position: { type: [string, 'null'] }
        phone: { type: [string, 'null'] }
        roles:
          type: array
          items: { enum: [participant, admin, superadmin] }
        profileComplete: { type: boolean }
    AuthResponse:
      type: object
      required: [user, csrfToken]
      properties:
        user: { $ref: '#/components/schemas/User' }
        csrfToken: { type: string }
    Event:
      allOf:
        - { $ref: '#/components/schemas/EventInput' }
        - type: object
          properties:
            id: { type: string, format: uuid }
            acceptsUploads: { type: boolean }
    EventInput:
      type: object
      required:
        - title
        - slug
        - shortCode
        - organizer
        - startsAt
        - endsAt
        - timezone
        - format
        - status
        - acceptUploadsFrom
        - acceptUploadsUntil
      properties:
        title: { type: string, maxLength: 300 }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        shortCode: { type: string, maxLength: 24 }
        description: { type: [string, 'null'] }
        organizer: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        timezone: { type: string }
        venue: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        format: { $ref: '#/components/schemas/EventFormat' }
        status: { $ref: '#/components/schemas/EventStatus' }
        tags: { type: array, items: { type: string } }
        acceptUploadsFrom: { type: string, format: date-time }
        acceptUploadsUntil: { type: string, format: date-time }
        maxFileSizeBytes: { type: integer, minimum: 1 }
        allowedMimeTypes: { type: array, items: { type: string } }
        blockedExtensions: { type: array, items: { type: string } }
        directAccessEnabled: { type: boolean }
    SubmissionInput:
      type: object
      properties:
        title: { type: [string, 'null'], maxLength: 300 }
        text: { type: [string, 'null'], maxLength: 50000 }
        link: { type: [string, 'null'], format: uri }
        hasFiles: { type: boolean }
    Submission:
      allOf:
        - { $ref: '#/components/schemas/SubmissionInput' }
        - type: object
          properties:
            id: { type: string, format: uuid }
            eventId: { type: string, format: uuid }
            userId: { type: string, format: uuid }
            status: { enum: [draft, processing, ready, failed, deleted] }
...
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: {}
  responses:
    Unauthorized:
      description: Необходима действительная Telegram-сессия
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Объект не найден или недоступен пользователю
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }